File validator completed

+++ src/main/java/Program/File/FileUploader.java
@@ -14,20 +14,21 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.Serializable;
-import java.nio.file.Files;
-import java.nio.file.Paths;
+import java.math.BigInteger;
 import java.security.DigestInputStream;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import java.util.List;
+import javax.annotation.PostConstruct;
 import javax.enterprise.context.SessionScoped;
 import javax.faces.application.FacesMessage;
 import javax.faces.context.FacesContext;
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.servlet.http.Part;
+import org.hibernate.Query;
 import org.hibernate.Session;
+import org.hibernate.exception.JDBCConnectionException;
 import org.primefaces.event.FileUploadEvent;
 import org.primefaces.model.UploadedFile;
 
@@ -50,8 +51,21 @@ public class FileUploader implements Serializable {
     
     private Part partFile;
     private UploadedFile uploadedFile;
+    private FileEntity holdingFile;
     
     private String startUpload;
+    
+    private String insertButtonValue;
+    private boolean showInsertButton;
+    private boolean disableInsertButton;
+    private boolean showRetryButton;
+    
+    @PostConstruct
+    public void init(){
+        insertButtonValue = "";
+        showInsertButton = false;
+        disableInsertButton = true;
+    }
 
     public HibernateUtil getHibernateUtil() {
         return hibernateUtil;
@@ -84,6 +98,46 @@ public class FileUploader implements Serializable {
     public void setStartUpload(String startUpload) {
         this.startUpload = startUpload;
     }
+
+    public FileEntity getHoldingFile() {
+        return holdingFile;
+    }
+
+    public void setHoldingFile(FileEntity holdingFile) {
+        this.holdingFile = holdingFile;
+    }
+
+    public String getInsertButtonValue() {
+        return insertButtonValue;
+    }
+
+    public void setInsertButtonValue(String insertButtonValue) {
+        this.insertButtonValue = insertButtonValue;
+    }
+
+    public boolean isDisableInsertButton() {
+        return disableInsertButton;
+    }
+
+    public void setDisableInsertButton(boolean disableInsertButton) {
+        this.disableInsertButton = disableInsertButton;
+    }
+
+    public boolean isShowInsertButton() {
+        return showInsertButton;
+    }
+
+    public void setShowInsertButton(boolean showInsertButton) {
+        this.showInsertButton = showInsertButton;
+    }
+
+    public boolean isShowRetryButton() {
+        return showRetryButton;
+    }
+
+    public void setShowRetryButton(boolean showRetryButton) {
+        this.showRetryButton = showRetryButton;
+    }
     
     public void startFileUpload(){
         this.startUpload = "File upload has started";
@@ -94,53 +148,147 @@ public class FileUploader implements Serializable {
      * 1) Validate file:
      *      - Make sure each line is of the same length
      *      - Compute checksum
-     *      - Compute length (both number of bytes and sequences)
+     *      - Compute size (both number of bytes and sequences)
      *      - Compare with existing file and determine if insert button should be a "Insert new file"/"Resume existing file"/"File already uploaded"
      * 2) Create/retrieve file reference from database to be used later for insertion of sequences
      * 3) 
      * @param event 
      */
     public void uploadFile(FileUploadEvent event){
-        uploadedFile = event.getFile();
+        /*uploadedFile = event.getFile();
+        
+        //1) Validate file
+        //- Make sure each line is of the same length
+        //- Compute MD5 checksum at the same time
+        //- Compute byte and sequence size at the same time
+        String md5Checksum = "";
+        int lineNum = 0;
+        long fileSize = uploadedFile.getSize();
+        */
         
         try {
-            //1) Validate file
-            //- Make sure each line is of the same length
-            //- Compute checksum at the same time
-            MessageDigest md = MessageDigest.getInstance("MD5");
+            /*MessageDigest md = MessageDigest.getInstance("MD5");
             InputStream is = uploadedFile.getInputstream();
             DigestInputStream dis = new DigestInputStream(is, md);
             
-            BufferedReader bReader = new BufferedReader(new InputStreamReader(is));
+            BufferedReader bReader = new BufferedReader(new InputStreamReader(dis));
             String line = new String();
-            long fileSize = uploadedFile.getSize(); 
+             
             int prevLineSize = 0;
-            int lineNum = 0;
             while((line=bReader.readLine())!=null){
                 lineNum++;
                 if(prevLineSize > 0 && prevLineSize != line.length()){ //last line is empty although gvim shows no last line
                     System.out.println(line);
-                    throw new RuntimeException("File "+uploadedFile.getFileName()+" does not contain equal length sequences!");
+                    throw new InvalidFileException("File "+uploadedFile.getFileName()+" does not contain equal length sequences!");
                 }
                 prevLineSize = line.length();
             }
+            byte[] digest = md.digest();// dis.getMessageDigest().digest();
+            for(int i=0; i<digest.length; i++){
+                System.out.print(digest[i]);
+            }
+            md5Checksum = String.format("%032x", new BigInteger(digest));
+            System.out.println("MD5 hash: "+md5Checksum);
             
-            byte[] digest = md.digest();
-            System.out.println(digest);
+            //initialize temp FileEntity set all variables to proceed with the remaining checks
+            holdingFile = this.createNewFile(uploadedFile.getFileName());
+            holdingFile.setBYTE_SIZE(fileSize);
+            holdingFile.setSEQUENCE_SIZE(lineNum);
+            holdingFile.setMD5_HASH(md5Checksum);*/
+            FileEntity checkedLengthAndChecksum = checkLengthAndComputeChecksum(event.getFile());
+            if(checkedLengthAndChecksum == null)
+                throw new InvalidFileException("File "+event.getFile().getFileName()+" does not contain equal length sequences!");
+            //Compare with existing file and determine if insert button should be a "Insert new file"/"Resume existing file"/"File already uploaded"
+            FileEntity existingFile = this.checkFileExists(hibernateUtil.getSession(), checkedLengthAndChecksum);
+            if(existingFile != null){
+                System.out.println("Existing file "+existingFile.getFILENAME()+" found");//debug
+                switch(existingFile.getSTATUS()){
+                    case INCOMPLETE :   setFacesMessage(FacesMessage.SEVERITY_INFO,
+                                            "File \""+existingFile.getFILENAME()+"\" has been uploaded before."
+                                            ,"");
+                                        this.insertButtonValue = "Resume upload for file \""+existingFile.getFILENAME()+"\"";
+                                        this.showInsertButton = true;
+                                        this.disableInsertButton = false;
+                                        break;
+                    case COMPLETED  :   setFacesMessage(FacesMessage.SEVERITY_INFO,
+                                            "Uploading has already completed for this file \""+existingFile.getFILENAME()+"\""
+                                            ,"");
+                                        this.showInsertButton = false;
+                                        this.disableInsertButton = true;
+                                        break;
+                    default         :   break;
+                }
+            } else {
+                System.out.println("No existing file found");//debug
+                this.insertButtonValue = "Insert new file \""+event.getFile().getFileName()+"\"";
+                this.showInsertButton = true;
+                this.disableInsertButton = false;
+            }
+            this.holdingFile = existingFile;
             
-        } catch (NoSuchAlgorithmException | IOException mdex) {
-            //Logger.getLogger(FileUploader.class.getName()).log(Level.SEVERE, null, ex);
-            FacesMessage msg = new FacesMessage(mdex.getMessage());
-            FacesContext.getCurrentInstance().addMessage(null, msg);
-        } catch (RuntimeException rex) {
-            FacesMessage msg = new FacesMessage(FacesMessage.SEVERITY_ERROR,rex.getMessage(),"");
-            FacesContext.getCurrentInstance().addMessage(null, msg);
-            
-            
+        }  catch (InvalidFileException ifex){
+            setFacesMessage(FacesMessage.SEVERITY_ERROR,ifex.getMessage(),"");
+            this.showInsertButton = false;
+        } catch(JDBCConnectionException jdbcex){
+            setFacesMessage(FacesMessage.SEVERITY_ERROR,"Database connection error!",jdbcex.getMessage());
+            this.showInsertButton = false;
+        } catch (Exception ex) {
+            setFacesMessage(FacesMessage.SEVERITY_ERROR,ex.getMessage(),"");
+            this.showInsertButton = false;
+        } 
+    }
+    /**
+     * Returns an initialized FileEntity object if it passes both length and checksum validation.
+     * Else returns null.
+     * <p>
+     * @param uploadedFile
+     * @return
+     * @throws Exception 
+     */
+    public FileEntity checkLengthAndComputeChecksum(UploadedFile uploadedFile) throws Exception{
+        FileEntity checkedFile;
+        String md5Checksum = "";
+        int lineNum = 0;
+        long fileSize = uploadedFile.getSize();
+        MessageDigest md = MessageDigest.getInstance("MD5");
+        InputStream is = uploadedFile.getInputstream();
+        DigestInputStream dis = new DigestInputStream(is, md);
+
+        BufferedReader bReader = new BufferedReader(new InputStreamReader(dis));
+        String line = new String();
+
+        int prevLineSize = 0;
+        while((line=bReader.readLine())!=null){
+            lineNum++;
+            if(prevLineSize > 0 && prevLineSize != line.length()){ //last line is empty although gvim shows no last line
+                System.out.println(line);
+                return null;
+            }
+            prevLineSize = line.length();
         }
+        byte[] digest = md.digest();// dis.getMessageDigest().digest();
+        for(int i=0; i<digest.length; i++){
+            System.out.print(digest[i]);
+        }
+        md5Checksum = String.format("%032x", new BigInteger(digest));
+        System.out.println("MD5 hash: "+md5Checksum);
+            
+        //initialize temp FileEntity set all variables to proceed with the remaining checks
+        checkedFile = this.createNewFile(uploadedFile.getFileName());
+        checkedFile.setBYTE_SIZE(fileSize);
+        checkedFile.setSEQUENCE_SIZE(lineNum);
+        checkedFile.setMD5_HASH(md5Checksum);
+        
+        return checkedFile;
+    }
+    
+    public void setFacesMessage(FacesMessage.Severity level, String headline, String description){
+        FacesMessage msg = new FacesMessage(level,headline,description);
+        FacesContext.getCurrentInstance().addMessage(null, msg);
     }
     
     public void insertFileAndSequences() {
+        System.out.println(holdingFile.getFILENAME());
         
         Session session = hibernateUtil.getSession();
 
@@ -215,13 +363,82 @@ public class FileUploader implements Serializable {
         return addSequence(file,sequence);
     }
     
+    /**
+     * Based on the following criteria, return an existing FileEntity or null if it doesn't fulfill ALL criteria:
+     * <p>
+     * <strong>1) Hash value</strong>
+     * <ul>
+     * <li>If no existing file with given hash value, return null</li>
+     * <li>If existing files with given hash value, proceed to the next check</li>
+     * </ul>
+     * <strong>2) File size</strong>
+     * <ul>
+     * <li>If no existing files with given file size, return null</li>
+     * <li>If existing files with given file size, proceed to the next check</li>
+     * </ul>
+     * <strong>3) Number of sequences</strong>
+     * <ul>
+     * <li>If no existing files with given number of sequences, return null</li>
+     * <li>If existing files with given file number of sequences, proceed to the next check</li>
+     * </ul>
+     * <strong>4) Filename</strong>
+     * <ul>
+     * <li>If no existing files with given filename, return null</li>
+     * <li>If existing files with given filename, proceed to the next check</li>
+     * </ul>
+     * <p>
+     * WARNING: This is not a 100% method, there is a chance that a file with different content 
+     * but with the same hash, size, number of sequences and name is returned. If multiple files 
+     * are found that fulfills all the above conditions as the given file, the first result will
+     * be returned.
+     * 
+     * @param session
+     * @param file
+     * @return 
+     */
     public FileEntity checkFileExists(Session session, FileEntity file){
-        return (FileEntity) session.get(file.getClass(),file.getFILENAME());
+        List<FileEntity> results;
+        //1) Hash value
+        String hashValueQuery = "SELECT file "
+                                + "FROM FileEntity file "
+                                + "WHERE "
+                                    + "file.MD5_HASH = '"+file.getMD5_HASH()+"' AND "
+                                    + "file.BYTE_SIZE = "+file.getBYTE_SIZE()+" AND "
+                                    + "file.SEQUENCE_SIZE = "+file.getSEQUENCE_SIZE()+" AND "
+                                    + "file.FILENAME = '"+file.getFILENAME()+ "'";
+        Query q = session.createQuery(hashValueQuery);
+        results = q.list();
+        if(results.size() <= 0 ){
+            return null;
+        }
+        else{
+            return results.get(0);
+        }
     }
     
-    public void clearFiles() throws IOException{
-        this.partFile.delete();
+    /**
+     * Cancels the entire process and clears all temp variables/storage.
+     * 
+     * @throws IOException 
+     */
+    public void cancel() throws IOException{
+        if(this.partFile!=null) this.partFile.delete();
+        this.partFile = null;
         this.uploadedFile = null;
+        this.holdingFile = null;
+        this.disableInsertButton = true;
+        this.showInsertButton = false;
+        
+        System.gc();
     }
     
+    public static String bytesToStringUTFCustom(byte[] bytes) {
+        char[] buffer = new char[bytes.length >> 1];
+        for (int i = 0; i < buffer.length; i++) {
+            int bpos = i << 1;
+            char c = (char) (((bytes[bpos] & 0x00FF) << 8) + (bytes[bpos + 1] & 0x00FF));
+            buffer[i] = c;
+        }
+        return new String(buffer);
+    }
 }
